+++ Source/Lib/Decoder/Codec/EbDecPicMgr.h
@@ -44,6 +44,8 @@ void generate_next_ref_frame_map(EbDecHandle *dec_handle_ptr);
 EbDecPicBuf *get_ref_frame_buf(EbDecHandle *dec_handle_ptr, const MvReferenceFrame ref_frame);
 void svt_setup_frame_buf_refs(EbDecHandle *dec_handle_ptr);
 
+ScaleFactors *get_ref_scale_factors(EbDecHandle *dec_handle_ptr, const MvReferenceFrame ref_frame);
+
 #ifdef __cplusplus
 }
 #endif