+++ Source/Lib/Decoder/Codec/EbObuParse.h
@@ -197,4 +197,11 @@ void svt_setup_motion_field(EbDecHandle *dec_handle);
 EbErrorType decode_obu(EbDecHandle *dec_handle_ptr, uint8_t *data, uint32_t data_size);
 EbErrorType decode_multiple_obu(EbDecHandle *dec_handle_ptr, uint8_t **data, size_t data_size);
 
+static INLINE int allow_intrabc(const EbDecHandle *dec_handle) {
+    return  (dec_handle->frame_header.frame_type == KEY_FRAME
+        || dec_handle->frame_header.frame_type == INTRA_ONLY_FRAME)
+        && dec_handle->seq_header.seq_force_screen_content_tools
+        && dec_handle->frame_header.allow_intrabc;
+}
+
 #endif  // EbDecObuParser_h