+++ CHANGELOG.md
@@ -22,6 +22,8 @@ Decoder
 - Post Processing Filters support
 - Compound Mode (InterInter & InterIntra) Tool support
 - Decoder Film Grain support
+- Decoder SuperRes support
+- Decoder Reference Frame Scaling support
 
 Build and Testing
 - Improve CI